Clean code

+++ AudioAnalyzer/Sources/audio-utils/cheby_win-lib/cheby_win.cpp
@@ -24,7 +24,7 @@ void cheby_win(float *out, int N, float atten) {
 	double M, n, sum = 0, max = 0;
 	double tg = pow(10, atten / 20);  /* 1/r term [2], 10^gamma [2] */
 	double x0 = cosh((1.0 / (N - 1))*acosh(tg));
-	M = (N - 1) / 2;
+	M = double((N - 1) / 2);
 	if (N % 2 == 0) M = M + 0.5; /* handle even length windows */
 	for (nn = 0; nn < (N / 2 + 1); nn++) {
 		n = nn - M;
@@ -32,10 +32,10 @@ void cheby_win(float *out, int N, float atten) {
 		for (i = 1; i <= M; i++) {
 			sum += cheby_poly(N - 1, x0*cos(pi*i / N))*cos(2.0*n*pi*i / N);
 		}
-		out[nn] = tg + 2 * sum;
+		out[nn] = float(tg + 2 * sum);
 		out[N - nn - 1] = out[nn];
 		if (out[nn] > max)max = out[nn];
 	}
-	for (nn = 0; nn < N; nn++) out[nn] /= max; /* normalise everything */
+	for (nn = 0; nn < N; nn++) out[nn] /= float(max); /* normalise everything */
 	return;
 }
